+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * json: disabled ALLOW_COERCION_OF_SCALARS, to make it stricter
   > thru JSONMapper.disable(MapperFeature.ALLOW_COERCION_OF_SCALARS)
   > previously, [""] will convert [null] if target type is List<Integer>
+* search: put detailed error in SearchException
 
 ### 7.10.7 (04/29/2022 - 05/12/2022)